+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 * Support for filtering servers within Admin CP to narrow down results by name, email, allocation, or defined fields.
 * Setup scripts (user, mail, env) now support argument flags for use in containers and other non-terminal environments.
 * New API endpoints for individual users to control their servers with at `/api/me/*`.
+* Typeahead support for owner email when adding a new server.
 
 ### Changed
 * Creating a user, server, or node now returns `HTTP/1.1 200` and a JSON element with the user/server/node's ID.